+++ test/integration/Strategies/EulerStrategy/constants.ts
@@ -1,0 +1,2 @@
+export const euler = "0x27182842E098f60e3D576794A5bFFb0777E025d3";
+export const eulerMarkets = "0x3520d5a913427E6F0D6A83E07ccD4A4da316e4d3";
