+++ test/integration/Strategies/MarginTradingStrategy/constants.ts
@@ -1,0 +1,1 @@
+export const kyberNetwork = "0x818E6FECD516Ecc3849DAf6845e3EC868087B755";
