+++ test/integration/Strategies/YearnStrategy/YearnStrategy.test.ts
@@ -5,10 +5,10 @@ import type { SignerWithAddress } from "@nomiclabs/hardhat-ethers/dist/src/signe
 import type { Vault } from "../../../../src/types/Vault";
 import type { ERC20 } from "../../../../src/types/ERC20";
 
-import { tokens, yearnRegistry } from "../../../common/mainnet";
+import { tokens } from "../../../common/mainnet";
 import { getTokens, expandToNDecimals, fundVault } from "../../../common/utils";
 import { marginTokenLiquidity, marginTokenMargin, leverage } from "../../../common/params";
-import { yearnFixture } from "../../../common/fixtures";
+import { yearnFixture } from "./fixture";
 
 import { YearnStrategy } from "../../../../src/types/YearnStrategy";
 import { Liquidator } from "../../../../src/types/Liquidator";
