+++ test/integration/Strategies/YearnStrategy/constants.ts
@@ -1,0 +1,1 @@
+export const yearnRegistry = "0x50c1a2eA0a861A967D9d0FFE2AE4012c2E053804";
